Unable to keep entry alive, so read_file_by_block does not quite work
