+++ src/archive_reader/mod.rs
@@ -1,8 +1,10 @@
 mod archive;
 mod entries;
 mod iter;
-mod reader;
 #[cfg(test)]
 mod tests;
+mod entry;
 
 pub use archive::*;
+pub use entries::Entries;
+pub use entry::Entry;
